move duplicated function to a helper file
