+++ editor/src/components/canvas/canvas-strategies/strategies/flex-resize-basic-strategy.tsx
@@ -46,6 +46,7 @@ import {
   resizeBoundingBox,
 } from './resize-helpers'
 import { FlexDirection } from '../../../inspector/common/css-utils'
+import { getElementDimensions } from './flex-resize-helpers'
 
 export function flexResizeBasicStrategy(
   canvasState: InteractionCanvasState,
@@ -286,40 +287,5 @@ export function resizeWidthHeight(
         height: boundingBox.height,
       })
     }
-  }
-}
-
-type ElementDimensions = {
-  width: number | null
-  height: number | null
-  flexBasis: number | null
-} | null
-
-const getElementDimensions = (metadata: ElementInstanceMetadata): ElementDimensions => {
-  const getOffsetPropValue = (
-    name: 'width' | 'height' | 'flexBasis',
-    attrs: PropsOrJSXAttributes,
-  ): number | null => {
-    return foldEither(
-      (_) => null,
-      (v) => v?.value ?? null,
-      getLayoutProperty(name, attrs, styleStringInArray),
-    )
-  }
-
-  if (isLeft(metadata.element)) {
-    return null
-  }
-  const { value } = metadata.element
-  if (!isJSXElement(value)) {
-    return null
-  }
-
-  const attrs = right(value.props)
-
-  return {
-    width: getOffsetPropValue('width', attrs),
-    height: getOffsetPropValue('height', attrs),
-    flexBasis: getOffsetPropValue('flexBasis', attrs),
   }
 }
